Create Sum of first x and last x integers in an array
The question given is : The program must accept N integers and an integer X as the input. The program must find the sum of the first X integers as S1 and then find the sum of the last X integers as S2. Finally, the program must print the sum of S1 and S2 as the output.
1) I have taken n integers (size of array)
2) I have taken the n number of array values
3) Then taken the x value
Now, I have implemented the code such that it satisfies the above condition in the question. Hope it helps!